+++ tests/testthat/1test_ECP_XRJulia.R
@@ -1,5 +1,5 @@
 library(convexjlr)
-context("Exponential Cone Programming")
+context("Exponential Cone Programming with XRJulia backend")
 
 ## The original Julia version
 
@@ -9,9 +9,9 @@ context("Exponential Cone Programming")
 # println(p.status)
 # x.value
 
-test_that("Results for example of exponential cone programming", {
+test_that("Results for example of exponential cone programming with XRjulia", {
     skip_on_cran()
-    convex_setup()
+    convex_setup(backend = "XRJulia")
 
     ## The R version with convexjl.R
 
@@ -21,7 +21,7 @@ test_that("Results for example of exponential cone programming", {
 
     ## The R version with XRJulia directly
 
-    ev <- XRJulia::RJulia()
+    ## ev <- XRJulia::RJulia()
     ev$Command("using Convex")
     ev$Command("x = Variable(4)")
     ev$Command("p = satisfy(norm(x) <= 100, exp(x[1]) <= 5, x[2] >= 7, geomean(x[3], x[4]) >= x[2])")