+++ tests/testthat/test_ECP_ECOS.R
@@ -1,5 +1,5 @@
 library(convexjlr)
-context("Exponential Cone Programming with JuliaCall")
+context("Exponential Cone Programming with JuliaCall in ECOS solver")
 
 ## The original Julia version
 
@@ -9,7 +9,7 @@ context("Exponential Cone Programming with JuliaCall")
 # println(p.status)
 # x.value
 
-test_that("Results for example of exponential cone programming with JuliaCall", {
+test_that("Results for example of exponential cone programming in ECOS", {
     skip_on_cran()
     convex_setup(backend = "JuliaCall")
 
@@ -17,7 +17,7 @@ test_that("Results for example of exponential cone programming with JuliaCall", 
 
     x <- Variable(4)
     p <- satisfy(norm(x) <= 100, exp(x[1]) <= 5, x[2] >= 7, geomean(x[3], x[4]) >= x[2])
-    cvx_optim(p, solver = "SCS")
+    cvx_optim(p, solver = "ECOS")
 
     ## The R version with XRJulia directly
 
@@ -26,11 +26,12 @@ test_that("Results for example of exponential cone programming with JuliaCall", 
     ## The R version with JuliaCall directly
     ev <- JuliaCall::julia_setup()
     ev$command("using Convex")
+    ev$command("using ECOS")
     ev$command("x = Variable(4)")
     ev$command("p = satisfy(norm(x) <= 100, exp(x[1]) <= 5, x[2] >= 7, geomean(x[3], x[4]) >= x[2])")
-    ev$command("solve!(p, SCSSolver())")
+    ev$command("solve!(p, ECOSSolver())")
 
     ## Compare the results
 
-    expect_equal(value(x), ev$eval("x.value")) ## , .get = TRUE))
+    expect_equal(value(x), ev$eval("x.value")) #, .get = TRUE))
 })